accommodate ruamel.yaml quirks, roundtrip finally works!

+++ python/alacs_test/__init__.py
@@ -1,9 +1,9 @@
-from collections.abc import Mapping, ItemsView
+from collections.abc import Mapping
 from io import BytesIO
 from time import perf_counter_ns
-from typing import NamedTuple, Any
+from typing import NamedTuple, Any, Self
 
-from alacs import Memory, File, Encoded, Comment, Text, List, Dict, Key
+from alacs import Memory, File, Encoded, Comment
 import alacs.yaml
 import ruamel.yaml
 from ruamel.yaml.comments import CommentedMap
@@ -115,36 +115,51 @@ class TimedRuamel:
         self.buffer = alacs.yaml.YAML()
         self.ruamel = ruamel.yaml.YAML(typ="rt")
         self.ruamel.indent(mapping=2, sequence=4, offset=2)
+        if self.preserves(b"comment", b"{}\n#comment"):
+            raise ValueError("ruamel fixed after empty bug")
+
+
+    def load(self, value: bytes | None) -> Any:
+        if value is not None:
+            self.buffer.seek(0)
+            self.buffer.truncate()
+            self.buffer.write(value)
+        self.buffer.seek(0)
+        return self.ruamel.load(self.buffer)
+
+    def dump(self, any: Any) -> Self:
+        self.buffer.seek(0)
+        self.buffer.truncate()
+        self.ruamel.dump(any, self.buffer)
+        return self
 
     def __bytes__(self) -> bytes:
         self.buffer.seek(0)
         return self.buffer.getvalue()
 
-    def _load_buffer(self) -> CommentedMap:
-        self.buffer.seek(0)
+    def preserves(self, mark: bytes, value: bytes) -> bool:
+        return mark in bytes(self.dump(self.load(value)))
+
+    def _load_file(self) -> CommentedMap:
         try:
             with self.load_timer:
-                result = self.ruamel.load(self.buffer)
+                result = self.load(None)
         except ScannerError:
-            self.buffer.seek(0)
-            print(self.buffer.getvalue().decode())
+            print(bytes(self).decode())
             raise
         if not isinstance(result, CommentedMap):
             raise AssertionError(f"expected CommentedMap, got: {type(result)}")
-        self.buffer.seek(0)
         return result
 
     def roundtrip(self, file: CommentedMap) -> CommentedMap:
-        self.buffer.seek(0)
-        self.buffer.truncate()
         with self.dump_timer:
-            self.ruamel.dump(file, self.buffer)
-        return self._load_buffer()
+            self.dump(file)
+        return self._load_file()
 
     def translate(self, file: File) -> CommentedMap:
         with self.alacs_timer:
             self.buffer.encode(file)
-        return self._load_buffer()
+        return self._load_file()
 
     def timers(self) -> None:
         print("   ruamel.yaml RT")
